Revert "Refactor/ Remove SampleClusterArray in favor of std::vector (#3617)"
This reverts commit be35ef768afba530e566fb9e00e8bc463630ac37.

+++ src/deluge/model/sample/sample_recorder.h
@@ -78,7 +78,7 @@ public:
 
 	uint32_t numSamplesExtraToCaptureAtEndSyncingWise;
 
-	size_t firstUnwrittenClusterIndex = 0;
+	int32_t firstUnwrittenClusterIndex = 0;
 
 	// Put things in valid state so if we get destructed before any recording, it's all ok
 	int32_t currentRecordClusterIndex = -1;
